requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# Anthropic Claude
anthropic>=0.40.0

# MCP (Model Context Protocol)
mcp>=1.20.0

# Google Calendar
google-auth>=2.36.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.154.0

# RAG / Vector Store
chromadb>=0.5.0
langchain>=0.3.0
langchain-community>=0.3.0
sentence-transformers>=3.3.0
pypdf>=5.1.0

# OpenAI (para embeddings)
openai>=1.57.0

# Utilities
httpx>=0.28.0
aiofiles>=24.1.0
python-multipart>=0.0.18

# Dev/Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=24.10.0
ruff>=0.8.0
